Fix for 1-string_nconcat.c

+++ more_malloc_free/1-string_nconcat.c
@@ -17,7 +17,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 		s1 = "";
 	if (s2 == NULL)
 		s2 = "";
-	if ((int) n > strlen(s2))
+	if (n > strlen(s2))
 		n = strlen(s2);
 	arr = malloc(strlen(s1) + n + 1);
 	if (arr == NULL)
